refactor(mespapiers): Move all search relative files into same folder
Since we want extract it later to use it on other app
