+++ packages/cozy-mespapiers-lib/src/components/Home/HomeLayout.jsx
@@ -7,7 +7,7 @@ import FeaturedPlaceholdersList from '../Placeholders/FeaturedPlaceholdersList'
 import { usePapersDefinitions } from '../Hooks/usePapersDefinitions'
 import { useMultiSelection } from '../Hooks/useMultiSelection'
 import { getFeaturedPlaceholders } from '../../helpers/findPlaceholders'
-import { useSearch } from '../Contexts/SearchProvider'
+import { useSearch } from '../Search/SearchProvider'
 import HomeToolbar from './HomeToolbar'
 import SearchHeader from './SearchHeader'
 import Content from './Content'
